CUDA-specific execution engine moves to core/cuda
This starts a deCUDAification effort for visible TC parts.  We want to
abstract high-level API from CUDA to allow for more backends.  This
commit moves CUDA-specific ExecutionEngine to a newly created core/cuda.
The following commits will move more related classes and introduce a
common abstract layer.
