+++ include/tc/aten/aten_compiler.h
@@ -22,7 +22,7 @@
 #include <ATen/ATen.h>
 #include <ATen/DLConvertor.h>
 
-#include "tc/core/execution_engine.h"
+#include "tc/core/cuda/cuda_execution_engine.h"
 #include "tc/core/utils/cuda_info.h"
 #include "tc/lang/parser.h"
